This program also finds a path from the start to the end location, but more efficiently.
The program scans the Grid.txt file to produce a 2 dimensional list with a start location and a end location. It will then find a path from the start to the end using either a Greedy Search or a A* Search. Once a path is found, it will create a path.txt file illustrating the path. This program was written in python using the anaconda IDE.

NOTE: The Grid.txt file must be in the same directory as the Uninformed search.py file for the program to function. This is where the path.txt file will be found.